Changed SpyChance type to float

+++ CISpies/Config.cs
@@ -9,7 +9,7 @@ namespace SCPPlugins.CISpies
         public bool IsEnabled { get; set; } = true;
         public bool Debug { get; set; }
         [Description("Chance to respawn a spy in a MTF wave (in percent)")]
-        public int SpyChance { get; set; } = 25;
+        public float SpyChance { get; set; } = 25;
         [Description("Specifies if a spy should spawn only once per round")]
         public bool SpawnSpyOnce { get; set; } = true;
         [Description("Value added to spy chance depending on modifier type")]
